pyproject.toml
[project]
name = "liveportrait"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations==1.4.10",
    "datasets>=3.6.0",
    "ffmpeg-python==0.2.0",
    "imageio==2.34.2",
    "imageio-ffmpeg==0.5.1",
    "ipykernel>=6.29.5",
    "lightning>=2.5.1.post0",
    "lmdb==1.4.1",
    "matplotlib==3.9.0",
    "numpy==1.26.4",
    "onnx==1.16.1",
    "onnxruntime-gpu==1.18.0",
    "opencv-python==4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=10.2.0",
    "pykalman==0.9.7",
    "pyyaml==6.0.1",
    "rich==13.7.1",
    "scikit-image==0.24.0",
    "scipy==1.13.1",
    "snac>=1.2.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm==4.66.4",
    "transformers==4.47.0",
    "tyro==0.8.5",
    "vector-quantize-pytorch>=1.22.16",
    "wandb>=0.19.10",
]
